fix(mobile): show pricing for bare/gateway model ids in the AI settings dropdown
Same root cause as the cost estimate: the dropdown's cost display used
getModelInfo directly, so 'deepseek-v4-flash' (an entry models.dev lists
with no pricing) showed nothing while 'deepseek/deepseek-v4-flash' carries
the price. Move the priced fuzzy lookup into models-dev (getPricedModelInfo)
and use it for both the dropdown trigger and rows (and the cost estimator).
Genuinely free models still render 'free' — that's correct.
